allow to import mixin as external nodejs module See https://github.com/microsoft/Oryx/issues/172 for more info about the issue
