+++ .gitignore
@@ -32,3 +32,5 @@ Desktop.ini
 # Mac crap
 .DS_Store
 
+# Visual Studio Code plugins
+.history
