+++ .travis.yml
@@ -10,6 +10,8 @@ node_js:
   - '8'
 before_script:
   - npm prune
+install:
+  - npm install
 after_success:
   - npm run semantic-release
 branches:
